Removed WindowsApplication and LinuxApplication
- Added Win32Common which is an application layer project and received the WindowsPlatform\Window class.  This is shared between WindowsApplication\GdiPlusRenderer and WindowsApplication\OpenGLRenderer.
- Had to add a hack in Window.cpp for IUnknown because combaseapi.h in the Windows 8.1 SDK isn't conformant with /permissive-.
- Removed WindowsPlatform and LinuxPlatform.  They were empty after the Window move.
- refs #48
