--- LinuxPlatform/BUILD
@@ -1,7 +1,0 @@
-cc_library(
-    name = "linux_platform",
-    visibility = ["//visibility:public"],
-
-    srcs = glob(["*.cpp"]),
-    hdrs = glob(["*.h"]),
-)
